\name{spLag-package}
\alias{spLag-package}
\alias{spLag}
\docType{package}
\title{
\packageTitle{spLag}
}
\description{
\packageDescription{spLag}
}
\details{
\packageDESCRIPTION{spLag}
\packageIndices{spLag}
}
\author{
\packageAuthor{spLag}

Maintainer: \packageMaintainer{spLag}
}
%\references{
%~~ Literature or other references for background information ~~
%}
%~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
%~~ documentation directory ~~
\keyword{ package }
%\seealso{
%~~ Optional links to other man pages, e.g. ~~
%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
%}
%\examples{
%~~ simple examples of the most important functions ~~
%}
